Add example deploying from an encapsulated ostree commit
Instead of calling from a remote repository, this example uses an
ostree commit constructed in a previous stage of the pipeline and
encapuslates it into an oci-archive. This oci-archive is then used
to deploy the OS. By doing this, we work around the problem of relying
on remote repositories in a testing context.