Create separate aleph stage
Refactor the code in the deploy stage to create a separate aleph
stage. A new stage prefered as both the org.osbuild.ostree.deploy
and org.osbuild.ostree.deploy.container stages need an aleph file.
Aleph files are also a coreos specific feature, which is why a
separate stage is beneficial.