osbuild/util/ostree: add `parse_deployment_option()`
Moved `default` option under `deployment` to
eliminate the need for an ostree specific top-level schema
option.

With this change, created a new `parse_deployment_option()`function which
retrieves the nested options under the  `deployment` object, reducing
code duplication for several stages. Additionally, several changes to the schema was was made
to restrict users from inputting `osname`, `ref`, and `serial` options
when `default` was set to true.